Move channel endpoints to new channels trait
