+++ src/Concerns/Api/StreamsTrait.php
@@ -104,37 +104,4 @@ trait StreamsTrait
 
         return $this->get('streams/markers', $parameters, $paginator);
     }
-
-    /**
-     * Gets channel information for users.
-     *
-     * @see https://dev.twitch.tv/docs/api/reference/#get-channel-information
-     *
-     * @param array<string, mixed> $parameters
-     *
-     * @return \romanzipp\Twitch\Result
-     */
-    public function getChannels(array $parameters = []): Result
-    {
-        $this->validateAnyRequired($parameters, ['broadcaster_id']);
-
-        return $this->get('channels', $parameters);
-    }
-
-    /**
-     * Modifies channel information for users.
-     *
-     * @see https://dev.twitch.tv/docs/api/reference/#modify-channel-information
-     *
-     * @param array<string, mixed> $parameters
-     * @param array<string, mixed> $body
-     *
-     * @return \romanzipp\Twitch\Result
-     */
-    public function updateChannels(array $parameters = [], array $body = []): Result
-    {
-        $this->validateAnyRequired($parameters, ['broadcaster_id']);
-
-        return $this->patch('channels', $parameters, null, $body);
-    }
 }
